create /src/types/posts for PostDataType
