+++ src/types/posts/index.ts
@@ -1,0 +1,10 @@
+export type PostDataType = {
+  id: string;
+  title: string;
+  create: string;
+  update: string;
+  tags: string[];
+  image: string;
+  content: string;
+  locale?: string;
+}
